Fixed cmake to work on my Arch Linux instance.
Had to to specify which version of OpenGL GL was wanted. This
happened because both a legacy version and the GLVND version is
installed. Had to explicitly include OpenGL in the packages.
Commented out the static GLEW switches as the distro is managing
the shared object files and header files. Also had to add the LIBRARIES
variables for OpenGL and GLEW. Had to explicitly include cmath otherwise
we get an undefined error on the function round.