Finished Assignment 01
Commented out the build of BasicRenderApp. Also put cmake into debug
builds. All the renaming requested is done. Changed the render loop to
draw a Box object. Though it just keeps going and it might segfault
once it reaches the edge of the screen. I created a box object to hold
my variables, I'm not that familiar with C++ so that took some
effort. Did some renaming in the render engine and stripped trailing
whitespace. Changed the drawing function to draw my Box object. The
Box object itself is setup. I don't know the proper ways to deinitial
objects in C++.